+++ src/Actions/Suma.java
@@ -2,7 +2,7 @@ package Actions;
 
 public class Suma extends Calculacion {
 
-    public Suma(int firstEntry, int secondEntry){
+    public Suma(float firstEntry, float secondEntry){
         calc(firstEntry, secondEntry);
     }
 
@@ -10,7 +10,7 @@ public class Suma extends Calculacion {
 
     }
 
-    public void calc(int s1, int s2){
+    public void calc(float s1, float s2){
         result = s1 + s2;
     }
 }
